Refactored controller into its own entity
This will make testing instructions possible: eg it will be possible to test
that the BRANCHZ instruction sets the PC_BRANCH signal with the correct
offset, but if Z is isn't set then PC_INCREMENT is set instead.

The existing top level tests can of course remain and be extended.