Reworking CPU/Control interface to use muxes
Instead of clocking out, for example, the ADDRESS from within the
controller, instead divert it from the PC, the left register (AKA operand) -
as used by STORER or DATA_IN - as used by STOREI using muxes.  Do similar
for the left (operand) ALU input, and the register write input. The muxes wire
things together in the cpu top level.

This is loosely based on ideas from reading about MIPS.

This allows a few states to be removed, fe. jumps/branches are always 3
(including fetch) from 5 (taken) or 3 (skipped). Further trimming is probably
possible.

The downside is it is a litle harder to follow what is going on from within the
controller.

Added a ALUI, immediate ALU ops. eg ADDI r0,0x123

STOREI and LOOADR are basically untested.

And this still needs tests. :(