Added a call stack
CALLJUMP, CALLBRANCH and CALLRETURN.

All three use the left *and* right register fields in the instruction to
set the stack pointer register.  LEFT=Address output to push/pop the PC,
RIGHT=the register to inc/dec.

Added inc and dec signals to the register entity, along with tests. This
saves use of the ALU, which would add another cycle.

The test fib program now uses 2 levels of call stack, and pretty much
every instruction. It now needs 64 words to play in. :)