Make cpu testbench reflect the FPGA setup
After running some improvements on the FPGA, I spotted some failing ops.
This was caused by the cpu testbench not reflecting the fact that the
FPGA memory was synchronus. Change the testbench to better match how the
external entity is rigged with mocked out video, LED and 7 segmenent
display, with a clock divider, address decoder and sync memory.

This required a new register, imaginativly called temporary, which holds
the address read by storei, since otherwise it would not be held across
clock ticks.

TODO: pointless testbench for the temporary register