Push and pop
Adds another mux to route the left index to the write index in the
registerfile. This is neceassary because pop must increment one
register and write to the other one. Pop also makes use of the
temporary register to hold the data read in the s_fetch2 cycle,
before it is written to the register in s_popquick1.

Next up, multiple register pushing and popping in one instruction
and two words.