Added push and pop multi
Can now push and pop multiple registers using one instruction
(the "quick" (single) versions is still useful so has been left)

Somewhat annoyingly the temporary register is exposed to the
controller as it holds the register bitmap. It could probably
just be a local signal or variable, like the stacked flags,
which would probably make more sense.

The instruction format is:

Opcode, Stack Pointer, Register to push/pop

A following immeiate word, with the low 8 bits used, holds a
bitmap of registers to push/pop.

The INC/DEC register operations have been moved to the left hand
operand and are (for now) fixed in the top level. It is now
possible to INC/DEC and WRITE in the same cycle, and POPQUICK
has been modified to do this, moving the INC/DEC and READ/WRITEs
into different cycles, so that eg. on a push the decrement
happens first.

Seems to work quite well. Cycle usage is not fab, because the
inc/dec has to come in its own cycle before or after the read or
write. Cycle usage is therefore:

PUSH/PULLQUICK: 2+1
PUSH/PULLMULTI: 2+(2*count)

The point where multi is faster is therefore 3 registers:

3*(2+1) = 9 vs 2+(3*2) = 8

Moved the READ cycles to being registered in the cpu test bench,
which will also be done in the FPGA version. This means the
temporary register is only needed when DATA_IN read needs to be
retained across different reads, as is the case with POPMULTI
where the register bitmap needs to be kept around.