Made memory interface allow byte wide operations
Internal CPU addresses are now byte wide, ie. the PC increments
by two when going to the next instruction. Likewise the INC/DEC
register operations needed for stacking.

Internally, all operands remain 16 bit; the ALU does not know or
care that the useful data is 8 bits. This works by sign or zero
extending byte wide memory reads, using two new bits in these
instructions: 9 is the byte/word flag (0=word), and 8 is the
sign/unsigned (0=unsigned).

A new entity, businterface, now wraps the cpu-proper and contains
the databus munging. As far as the cpu entity goes, it deals with
byte wide addresses that are a word wide, with byte values right
aligned. The businterface moves these byte accesses to the correct
word half and sets the upper/lower signals appropriately so that
reads and writes can be made (in the testbench, or later in the
FPGA memory or external memory) on the needed half, which is
vital for writes so that the other half is not modified.

LOADI supports byte wide reads, though it is pointless and may
be removed later. Since immediates are always in subsequent words
then it makes no sense to do extensions when the assembler can
just insert the right word value.

There are two other, probably better, arrangements for the bus-
interface:

* Move the combination logic in its own entity into the cpu entity
* Make it an in-line entity next to the cpu instead of putting the
cpu inside, or putting the businterface inside.

I'll probably switch it over to the first one.