Added inital assembler definition using customasm
customasm (https://github.com/hlorenzi/customasm) is a data driven assembler
for arbitary processors.

The assembler is not perfect; it lacks macros and other things, but it is
very easy to define the nmominics and their opcodes, different addressing
modes, register sets etc.

Previously I was uasing axasm (https://github.com/wd5gnr/axasm) but this
looks better. :)

It's now possible to write decent looking code:

store.b (r1,0x1234),r0

Instead of:

storerdb r1,r0,0x1234

Likewise the ALU operations just have one name, and the assembler figures
out the opcode based on wether a register or an immediate is specified as
the oerand. And the ALU operations just need to be defined once. :)