Rework control to use memory clocked with the processor
Previously memory (in the main testbench and the FPGA) was clocked with
the external clock. In practise for the processor to be useful it needs
to use the same clock as the processor.

To fascilitate this the main and delayed clocks are now outputted
from the cpu entity so they can be used for clocking memory.

In the testbench, the addressdecoder has been moved inline with the
read and write actions. This was needed because the decoded outputs
were clocked. It's kind of reduandant having this external to thew
where those inputs are needed anyway.