Added arithmetic shifts and overflow flag
Based on the 68K behavior for airthmetic (signed) shifts: right will
pin the sign bit, and left will set the overflow flag if the sign
toggles but is otherwise identical to the logical shift.

Overflow is set on add, addc, sub and subc if the sign of the result
is wrong for signed operations. I *think* the logic is right, but it
may need tweeking.

shiftleft/right has been renamed to logicleft/right. The new
operations are arithleft/right in the assembler.