Removed clock from the ALU
This saves cycles in ALU ops, making them as fast as a clear; I'll
probably move clear into the ALU as a no-operand operation in a
later commit.

This means all states (except FETCH2) are using the memory bus, which
is a nice improvement.

The slowest instruction at the moment is CALLBRANCH/JUMP which has
to: dec the SP, write the PC, and read and jump/branch the subroutine.

Untested in the FPGA at present.