Added a reset vector at location 0
This is a warm up to looking at interrupts, which will need vectors.
Snake and the fib program code have been adjusted to add a reset
vector.

Improved HALT by making use it of properly in the test bench to end
the simulation, and fixed the cpudef line for it.